Move transaction validation to new package
