+++ common/src/main/scala/co/topl/modifier/transaction/validation/package.scala
@@ -1,6 +1,6 @@
-package co.topl.modifier
+package co.topl.modifier.transaction
 
-package object transaction {
+package object validation {
 
   object implicits
       extends SyntacticallyValidatable.ToSyntacticallyValidatableOps
